refactoring authentication from es5 to es6
