+++ middleware/auth.middleware.js
@@ -8,7 +8,6 @@ module.exports = (req, res, next) => {
 
   try {
     //строку Bearer token помещаем в массив и забираем token
-
     const token = req.headers.authorization.split(" ")[1].trim();
     if (!token) {
       res.status(401).json({ message: "Нет авторизации" });
