+++ routes/auth.routes.js
@@ -1,86 +1,8 @@
-const { Router } = require("express");
-const bcrypt = require("bcryptjs");
-const { check, validationResult } = require("express-validator");
-const jwt = require("jsonwebtoken");
-const config = require("config");
-const User = require("../models/User");
-const router = Router();
+import express from "express";
+const router = express.Router();
+import { signup, signin } from "../controllers/auth.js";
 
-router.post(
-  "/register",
-  [
-    check("email", "Некорректный email").isEmail(),
-    check("password", "Минимальная длина пароля 6 символов").isLength({
-      min: 6,
-    }),
-  ],
-  async (req, res) => {
-    try {
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return res.status(400).json({
-          errors: errors.array(),
-          message: "Некорректные данные при регистрации",
-        });
-      }
-      const { email, password } = req.body;
+router.post("/register", signup);
+router.post("/login", signin);
 
-      const candidate = await User.findOne({ email });
-
-      if (candidate) {
-        return res
-          .status(400)
-          .json({ message: "Такой пользователь уже существует" });
-      }
-
-      const hashedPassword = await bcrypt.hash(password, 12);
-      const user = new User({ email, password: hashedPassword });
-      await user.save();
-      res.status(201).json({ message: "Пользователь создан" });
-    } catch (e) {
-      res.status(500).json({ message: "Что-то пошло не так." });
-    }
-  }
-);
-
-router.post(
-  "/login",
-  [
-    check("email", "Некорректный email").normalizeEmail().isEmail(),
-    check("password", "Минимальная длина пароля 6 символов").isLength({
-      min: 6,
-    }),
-  ],
-  async (req, res) => {
-    try {
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return res.status(400).json({
-          errors: errors.array(),
-          message: "Некорректные данные при входе в приложение",
-        });
-      }
-
-      const { email, password } = req.body;
-      const user = await User.findOne({ email });
-      if (!user) {
-        return res.status(400).json({ message: "Пользователь не найден" });
-      }
-
-      const isMatch = await bcrypt.compare(password, user.password);
-      if (!isMatch) {
-        return res.status(400).json({ message: "Не верный пароль" });
-      }
-
-      const token = jwt.sign({ userId: user.id }, config.get("secretJwt"), {
-        expiresIn: "1h",
-      });
-
-      res.json({ token, userId: user.id });
-    } catch (e) {
-      res.status(500).json({ message: "Что-то пошло не так." });
-    }
-  }
-);
-
-module.exports = router;
+export default router;
